418: Refactored pagination with items per page
