+++ config/packages/api_platform.yaml
@@ -8,7 +8,6 @@ api_platform:
 
     docs_formats:
         jsonld: ['application/ld+json']
-        jsonopenapi: ['application/vnd.openapi+json']
         html: ['text/html']
 
     mapping:
