+++ src/Api/State/EventRepresentationProvider.php
@@ -6,7 +6,7 @@ use ApiPlatform\Metadata\CollectionOperationInterface;
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\State\ProviderInterface;
 use App\Model\IndexNames;
-use App\Service\ElasticSearchPaginator;
+use App\Service\ElasticSearch\ElasticSearchPaginator;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
 
@@ -19,15 +19,13 @@ final class EventRepresentationProvider extends AbstractProvider implements Prov
      */
     public function provide(Operation $operation, array $uriVariables = [], array $context = []): ElasticSearchPaginator|array|null
     {
-        // @TODO: should we create enum with 5,10,15,20
-        // Get page size from context.
-
         if ($operation instanceof CollectionOperationInterface) {
             $filters = $this->getFilters($operation, $context);
             $offset = $this->calculatePageOffset($context);
-            $results = $this->index->getAll(IndexNames::Events->value, $filters, $offset, self::PAGE_SIZE);
+            $limit = $this->getImagesPerPage($context);
+            $results = $this->index->getAll(IndexNames::Events->value, $filters, $offset, $limit);
 
-            return new ElasticSearchPaginator($results, self::PAGE_SIZE, $offset);
+            return new ElasticSearchPaginator($results, $limit, $offset);
         }
 
         return [$this->index->get(IndexNames::Events->value, $uriVariables['id'])['_source']];
