+++ src/Api/State/OrganizationRepresentationProvider.php
@@ -6,7 +6,7 @@ use ApiPlatform\Metadata\CollectionOperationInterface;
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\State\ProviderInterface;
 use App\Model\IndexNames;
-use App\Service\ElasticSearchPaginator;
+use App\Service\ElasticSearch\ElasticSearchPaginator;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
 
@@ -22,9 +22,10 @@ final class OrganizationRepresentationProvider extends AbstractProvider implemen
         if ($operation instanceof CollectionOperationInterface) {
             $filters = $this->getFilters($operation, $context);
             $offset = $this->calculatePageOffset($context);
-            $results = $this->index->getAll(IndexNames::Organization->value, $filters, $offset, self::PAGE_SIZE);
+            $limit = $this->getImagesPerPage($context);
+            $results = $this->index->getAll(IndexNames::Organization->value, $filters, $offset, $limit);
 
-            return new ElasticSearchPaginator($results, self::PAGE_SIZE, $offset);
+            return new ElasticSearchPaginator($results, $limit, $offset);
         }
 
         return [$this->index->get(IndexNames::Organization->value, $uriVariables['id'])['_source']];
